pyproject.toml
[project]
name = "mlflow-community-helm-charts"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "markdown",
    "pymdown-extensions",
    "pygments",
    "PyYAML",
]
